Reapply "split out ColorComponents implementation to fix Android build"
This reverts commit 9f30b9482162ee969537017856d25e9abce87d1c.
