+++ packages/react-native/ReactCommon/react/renderer/graphics/ColorComponents.cpp
@@ -1,0 +1,15 @@
+#include "ColorComponents.h"
+
+namespace facebook::react {
+
+static ColorSpace defaultColorSpace = ColorSpace::sRGB;
+
+ColorSpace getDefaultColorSpace() {
+  return defaultColorSpace;
+}
+
+void setDefaultColorSpace(ColorSpace newColorSpace) {
+  defaultColorSpace = newColorSpace;
+}
+
+} // namespace facebook::react
